+++ src/components/game/description/description.css
