+++ src/components/game/result/result.css
